gordon-packages includes only 'necessary' packages
Now it only includes things I expect to use all the time.
Things like auctex, dsvn, and csharp-mode can be manually installed when
necessary. mode_customizations.el ensures that not having them installed
won't cause any issues.

Note that git-commit-mode is a dependency of magit, so it doesn't need
to be explicitly installed.